Add types, move STrack into a new file
